requirements.txt
amqp==5.2.0
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
async-timeout==4.0.3
beautifulsoup4==4.12.3
billiard==4.2.0
cachetools==5.3.3
celery==5.3.6
certifi==2023.7.22
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
crispy-bootstrap5==2023.10
cryptography==41.0.5
defusedxml==0.7.1
distro==1.9.0
dj-database-url==2.2.0
dj-rest-auth==6.0.0
Django==4.2.7
django-allauth==0.58.1
django-appconf==1.0.6
django-autocomplete-light==3.9.7
django-compressor==4.4
django-countries==7.6.1
django-crispy-forms==2.1
django-easy-audit==1.3.5
django-filter==23.5
django-haystack==3.2.1
django-sass-processor==1.4
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
elastic-transport==8.12.0
elasticsearch==2.4.1
exceptiongroup==1.2.0
geographiclib==2.0
geopy==2.4.0
google-api-core==2.19.0
google-api-python-client==2.130.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.0
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.26.0
idna==2.10
kombu==5.3.5
numpy==1.26.4
oauthlib==3.2.2
openai==1.11.1
packaging==24.1
Pillow==10.1.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycountry==24.6.1
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
pydoodle==1.1.3
pyjet==1.9.0
PyJWT==2.8.0
pyparsing==3.1.2
pypdf==4.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
python3-openid==3.2.0
rcssmin==1.1.1
redis==5.0.2
requests==2.25.1
requests-oauthlib==1.3.1
rjsmin==1.2.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
soupsieve==2.5
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.18
vine==5.1.0
wcwidth==0.2.13
